requirements.txt
torch>=2.0.0
sentencepiece>=0.1.99
sacrebleu>=2.3.1
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
tokenizers>=0.15.0
datasets>=2.14.0
